Allow length of Service Name to be specified
Allows length of service name within MultiEPG to be specified depending on layout of skin.

Min of 5, max of 20 and defaults to 7 (matching existing hard-coded length).

This option only applies to MultiEPG because by default, depending on skin, service name may be cut off.

Thanks to @seagen for noticing issue. More details mentioned in following thread:

https://www.world-of-satellite.com/showthread.php?62653-New-EPG-files-User-trial-and-feedback&p=497546&viewfull=1#post497546