Detect split programs when setting timers (#543)
Many commercial channels split longer programmes, usually films, with short news bulletins. This change adds an option to detect this and automatically extend the recording across the news so you don't accidentally miss the second half. Most places in the UI support now this:

    2nd infobar
    Channel List
    EPGs
    Event View

A new setting, split_programme_minutes has been added, which defaults to 15 minutes, so should cover both the entertainment style news and the longer bulletins seen on ITV from time to time. We can reduce the default or have it switched off by default if necessary.

The detection is uses the name and short description of the events. This seems to work well on the UK's ITV, 5, Sony and Paramount channels.