[Geolocation.py]  Allow config control of geolocation (#167)
This change adds a config entry "config.misc.enableGeolocation" to control if geolocation is run at startup.  The default is to allow it to run once every boot until the geolocation lookup succeeds and then it will never run again.  It can be enabled for one more run by setting "config.misc.enableGeolocation" to True.  It should be noted that the previous code would not allow geolocation to run more than once in any boot session.  This case is now clearly stated with an appropriate log message.

In the future "config.misc.enableGeolocation" will not be forced to False on a successful lookup but will allow the user to specifically control if geolocation is permitted to run more frequently.