[Timezones.py] Update to match changed geolocation
This change accounts for the change in geolocation execution in that geolocation can now only run once on the first start. This allows the initial timezone selection to be suggested based on the geolocation data. If geolocation is unavailable or if a proxy is detected then the image based defaults will continue to be used.

This change enforces that geolocation can only provide a time zone suggestion once in the first run wizard.

(cherry picked from commit 570610b8f15248ca78fcaa770f770d18e907a110)