[estring/event]  removing character ';'/ add items field
- [estring.cpp] removing character ';' , a string is truncated without
an invalid character at the end of line.
more info:
https://github.com/openatv/enigma2/issues/1681
-- [event.cpp] If you look at "ETSI EN 300 468 V1.16.1 (2019-08)" in the
"6.2.15 Extended event descriptor" section, there are two types of text
fields - items and text. The items field consists of one or more
item_description and item pairs. Tricolor transmits information about
the country, year. actors, etc. in the items field, and the description
of the event in the text field, which is displayed in enigma2, while the
items field is ignored by the enigma. This is easy to see if you install
dvbsnoop and capture several Event Information Table fields.
To do this, you need:
1. Install dvbsnoop
opkg update
opkg install dvbsnoop

2. Capture 10 EIT fields
dvbsnoop -N 10 0x12> /tmp/info.txt