[InputDevice.py] Add button remapping facility
This code uses a companion component in the C++ RCinput code to allow specific remote controls to have buttons remapped.
Sister commit: https://github.com/OpenViX/enigma2/commit/988a888f2a6fc51ba662d85a04141de56f58f128

The rcpositions.xml file for a remote control can have a "keyid=" attribute that nominates the key id transmitted by the remote control and a "remap=" attribute that nominates the key id value that should be send to Enigma2.  The original keyid value is no longer seen by Enigma2.  It is if the remote control hardware/drivers have been changed to emit the remapped code.  For example, the gb4 remote control for the GigaBlue UHD UE 4K could have the following line in rcpositions.xml:
    <button keyid="KEY_F1" remap="KEY_HELP" name="HELP" title="HELP" pos="26,409" shape="rect" size="28,12" />
This takes the KEY_F1 code sent by the remote control and sends a KEY_HELP code to Enigma2.  That is, this GigaBlue remote control now has a HELP button for the loss of the F1 button.