[config.py] Allow non string values in ConfigSelection
This commit allows non-string values to be used in ConfigSelections, e.g. booleans and integers. These values can now be safely saved to the settings file. On reading the settings file the correct data type is retrieved.

This code ensures the correct functioning of the many instances of ConfigSelection that are already using non-string values, and offers greater flexibility to developers writing new code.

Commit author: PRL