[ConfigList.py] Make RED/GREEN control optional
This change allows the calling code to specify if the RED/GREEN, cancel/save, options should be controlled by this module.   The argument to control this actions if a Boolean "fullUI".  If set to False (the default) the code to operate as it did before the Setup/ConfigList refactor was started.  If set to True the code will offer a standardised UI to offer the RED/GREEN, cancel/save UI elements.

The fullUI argument is optional and if not specified is False.  Currently the new Setup.py uses this argument but it is now available to any coders who would like to use a standardised Setup based UI.