[(Power)TimerEntry] Clean up TimerLog
[TimerEntry]

Add help to TimerLog.

Remove the explicit updating of the log description entry and replace
it by using the onSelectionChanged callbacks in MenuList.

Simplify the navigation methods by calling the navigation methods
in MenuList instead of the methods in MenuList's instance.

Rename those methods to match those in MenuList, but retain the old
methods and mark them as deprecated.

Use assignment to the self["logentry"].text property instead of
directly calling self["logentry"].setText().

[PowerTimerEntry]

Remove the body of TimerLog and replace it by inheriting from
TimerEntry.TimerLog.

This has the side-effect of making the user's time/date settings
work in PowerTimerEntry.TimerLog.