[config.py] ConfigBoolean trueValues
Some plugins, e.g. MediaPortal, inherit ConfigBoolean but fail to call ConfigBoolean.__init__(). This means the plugin fails to read self.trueValues and ends up crashing.

This is a temporary workaround only. The plugin needs fixing. If not it should be removed from the feeds. But that will probably cause a lot of complaining.

From the plugin:

class ConfigBooleanExt(ConfigBoolean):
 def __init__(self, default = False, descriptions = boolean_descriptions):
  ConfigElement.__init__(self)

This should be:

class ConfigBooleanExt(ConfigBoolean):
 def __init__(self, default = False, descriptions = boolean_descriptions):
  ConfigBoolean.__init__(self)