[Wizard] Fix manipulation of ConfigLists
In Wizard.afterAsyncCode(), when a <config/> has type="configList",
the contents of self.configInstance["config"].list are copied into
self["config"], the contents of self.configInstance["config"] are
cleared (as if by a close()), and self["config"].list is assigned
to self.configInstance["config"].list.

This means the setup screen's config list is displayed as if it was
the Wizard's.

However, when the config lists were copied, it was done by doing
self["config"].l.setList(self.configInstance["config"].list)

This only set the ConfigList instance's copy of the config list,
and not the one in ConfigList itself.

This could have some odd side-effects, especially in Screens.Time
when it is run in the startup wizard.

All instances of self["config"].l.setList(...) have been changed
to self["config"].list = ... so that both the list in ConfigList
and its instance are kept consistent, and the cleaner,
more compact property interface is used.

Also removed the call to self.configInstance.setAnimationMode(0)
in this area of the code - again, it's an around-the-back call into
the instance state, and self.configInstance is hidden anyway, so
it has no effect.

Added a short comment about what this section of code is doing (most
of the introductory paragraph of the commit note is taken from that
comment).