[MessageBox] autoResize(): dynamic lower text margin
Previously there was a fixed margin of 25px between the text area and the list area which sometimes was not sufficient. Now, if the text area is wider than 520 the bottom margin will increase at the same ratio. e.g. 520 will give a margin of 30, whereas 780 will have a margin of 45.