Revert "[UsageConfig] Change defaut log format"
This reverts commit ea5441d6b7e396352ba9bfbaa7d4d6154f1a4798.

I said don't change this. I added this feature because people asked for it as an option. And now they are to lazy to turn it on, so it is being forced on the rest of us who don't want it. This takes up unnecessary space at the start of every single line of the log and force sideways scrolling to be able to read the average length log line.