filepush: work around bug in OE Zeus glibc. Fix recordings that hang enigma.
On MIPS receivers the default signal blocking mask for a thread appears
to be undefined (and it's really random). If we don't set it explicitly,
it will sometimes let SIGUSR1 work and on other times it won't.

So set the signal blocking mask in the recording thread explicitly now,
and that seems to work.

On ARM receivers this issue never showed.

Also moved installing the signal handler from the thread to the enigma
parent process, because signal handlers are global for the process.