[Tools.Directories] SCOPE_FONT: Check the "fonts" folder (#2852)
Before this commit it was necessary to use a path when accessing a font from the current skin. It makes sense, if we are trying to locate a font, that the "fonts" folder is tested directly as this is the most likely place a font file will be located in the currently active skin.

e.g. to access...
<font name="Regular" filename="/usr/share/enigma2/<skin-name>/fonts/SFNSDisplay-Regular.otf" scale="100" />
or this...
<font name="Regular" filename="fonts/SFNSDisplay-Regular.otf" scale="100" />

it is now possible to do this...
<font name="Regular" filename="SFNSDisplay-Regular.otf" scale="100" />

It is wrong that it is possible to access a font located in "/usr/share/fonts" without using a path, but if accessing a font in "/usr/share/enigma2/<skin-name>/fonts" a path is required. This commit allows consistence between these locations.

Consistency needs to come first. We shouldn't have confusing traps like this in enigma.

Also remove some comments that if ever implemented would wreak havoc throughout the enigma community.

(cherry picked from commit 5db0dbf2262bf84fa94c27b39093d9851ab2bda6)