Fix font and itemHeight scale in TemplatedMultiContent
If in skin uses "*f" to use getSkinFactor on FHD skin getSkinFactor return 1.5 which is flaot.
Therefore, when it multiplied by the values on eval they are also converted to float.
To fix this allows use float in gFont and convert the value to int in setItemHeight.

(cherry picked from commit 3cbcef832168a5f309b2f0cf3840beebbeb9ee93)