loadSVG: allow scale svg only from height or width size
This allows to scale images if only height or width size is given.
In this case, the scale for width and height is calculated taking into account only one of the given dimensions size if they differ from the original size and the other dimension size is not given.

(cherry picked from commit 99f4b2677a72f46d8306149197858b6984e3a2db)