[config.py] ConfigClock: allow the default to be a tuple
The default argument for ConfigClock is a timestamp. This commit allows that argument to be either a timestamp or an hours/minutes tuple, e.g. 20:00 would be (20, 0) allowing the values to be in a human readable form rather than needing a calculator.

Integration directly in the class suggested by Simon