[ChoiceBox] Fix problem of "list" being used as a variable name in the argument list of this class
In Py3 function "zip" returns an iterator, not a list like Py2, so needs to be wrapped with a list but it is impossible to do that as the "list" token has been overwritten by the clas arguments.

So move these to another "def" where the "list" token is available.