[Satconfig] Tweak "Extra motor options"
When opening Satconfig, if any of the "Extra motor options" are not at their default settings the "Extra motor options" list will automatically be expanded. Otherwise it will remain contracted unless the user opens it.

Also skip this item when checking "isChanged()" as it is just a GUI item and is not saved to the tuner configuration..