[Renderer.Pixmap] Stop CHANGED_CLEAR clearing the pixmap inappropriately
This corrects an earlier fix to Renderer.Pixmap in commit 414da5c.

The instance pixmap should only be cleared on CHANGE_CLEAR if the
pixmap is being taked from renderer.source's pixmap atrtribute (when
it has one).

This fixes problems seen in the display of the tuner type icon in
the infobar in the iSkin-* and Q-FHD_BLACK infobar skins, where the
icon would be displayed after a reboot (or GUI restart), but disappear
after the first channel change.