[BackupManager/ImageManager] maintain list position on delete
Feature request by @ccs: After a backup has been deleted the selection highlight jumps to the first item in the list.
If you were in page 2 of the list (or even half way down page 1), there are a lot of keypresses needed to get you back to where you want to be.

This change maintains the selction highlight at the same position in the list, similar to how MovieSelection works on delete.