[InfoBarGenerics.py] Comments related to opening AutoTimer
Using "long green" to open AutoTimerOverview... it is required the the translation of "Auto Timers" is identical in both AutoTimer.po and enigma.po.
This also adds a test in the case where in enigma.po that phrase is trranslated but in AutoTimer.po it is not.

Related commit: https://github.com/OpenViX/enigma2/commit/b5d0879b1b530999d6ac79de9893b9cd6f68551f