[ImageManager] allow feed url format to be flexible
Previously the url format was rigid e.g. http://domain.com/json/machine because that was the format of the major images.

Now we use a pattern for the url and the machine name is inserted. This allows the machine name to be part of the query string if desired.

Also, stop using ConfigText for feeds URLs as these are not user configurable items.